requirements.txt
flask
requests
transformers
sentencepiece
openai-whisper
ffmpeg-python
